Adding new components
I added Tables which displays json data with the help of DisplayTables and JsonEdit. Also I added ForgotPassword component but couldn't work on it.  Other than that in my work in adding edit and delete button I tried a different approach through Table.js component and other two helping which are: ReadOnlyRow.js and EditableRow.js. This part is functional when used separately but I had a problem with integrating it into the project.